% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/bootstrap_phylo_sig.R
\name{bootstrap_phylo_sig}
\alias{bootstrap_phylo_sig}
\title{Gets \code{tree} phylogenetic signal from trait to \code{matrix}.}
\usage{
bootstrap_phylo_sig(trees, traits, sample, rep)
}
\arguments{
\item{trees}{A \code{list} object with nexus tree.}

\item{traits}{A \code{matrix} with traits.}

\item{sample}{A \code{numeric} integer. Is the sample size.}

\item{rep}{A \code{numeric} integer. Is the sample size.}
}
\description{
Gets \code{tree} phylogenetic signal from trait to \code{matrix}.
}
\examples{
my_result <- foo(iris)

}

