% Generated by roxygen2 (4.0.1): do not edit by hand
\name{correlogram.phylo4d}
\alias{correlogram.phylo4d}
\alias{spline.correlogram.phylo4d}
\title{Phylogenetic correlograms}
\usage{
correlogram.phylo4d(p4d, trait = names(tdata(p4d)),
  dist.phylo = "patristic", increment = 2, reps = 1000)

spline.correlogram.phylo4d(p4d, trait = names(tdata(p4d)),
  dist.phylo = "patristic", df = NULL, reps = 200, npoints = 100,
  max.it = 25)
}
\arguments{
\item{p4d}{a phylo4d object.}

\item{trait}{the trait(s) in the p4d object to use for correlogram estimation.
Can be a character vector giving the name of the trait(s) or a single number giving the column index
in the table of the data slot of the p4d object.}

\item{dist.phylo}{a character string specifying the method used to compute phylogenetic distances.
See Details.}

\item{increment}{a numeric value. Increment parameter for the uniformly distributed distance classes.}

\item{reps}{a numeric value. Number of repetitions for the estimation of p.values with randomization.}

\item{p4d}{a phylo4d object.}

\item{trait}{the trait(s) in the p4d object to use for correlogram estimation.
Can be a character vector giving the name of the trait(s) or a single number giving the column index
in the table of the data slot of the p4d object.}

\item{dist.phylo}{a character string specifying the method used to compute phylogenetic distances.
See Details.}

\item{df}{numeric. Degrees of freedom for the spline. Default is \code{sqrt(n)}.}

\item{npoints}{the number of points at which to save the value for the spline function
(and confidence envelope).}

\item{max.it}{a numeric. The maximum iteration for the Newton method used to estimate the intercepts.}

\item{reps}{a numeric value. Number of repetitions for the bootstraps.}
}
\value{
an object of class \code{correlog}.

an object of class \code{spline.correlog}.
}
\description{
This function estimate phylogenetic correlograms for one trait
(univariate Moran's I correlogram) or several (multivariate Mantel correlogram).

This function estimate phylogenetic spline correlograms for one trait
(univariate Moran's I spline correlogram) or several (multivariate Mantel spline correlogram).
}
\author{
This is a modified version of the \code{correlog}
function in the \pkg{ncf} package written by Ottar N. Bjornstad.

This is a modified version of the \code{spline.correlog}
function in the \pkg{ncf} package written by Ottar N. Bjornstad.
}
\references{
Bjornstad, O.N. & Falck, W. (2001)
Nonparametric spatial covariance functions: estimation and testing.
Environmental and Ecological Statistics, 8:53-70.
}

