% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utilities.R
\name{descentsNames}
\alias{descentsNames}
\title{Names of descents}
\usage{
descentsNames(phy, node)
}
\arguments{
\item{phy}{an object of class "\code{phylo}.}

\item{node}{the ID of the node.}
}
\value{
A vector tip labels.
}
\description{
This function return all the names of the descents (tips) of a given node.
}
