% Generated by roxygen2 (4.0.1): do not edit by hand
\name{directDescents}
\alias{directDescents}
\title{ID of direct descents}
\usage{
directDescents(phy, node, as.phylo = FALSE)
}
\arguments{
\item{phy}{an object of class "\code{phylo}.}

\item{node}{the ID of the node.}
}
\value{
A vector of node numbers.
}
\description{
This function return the node number (ID) of the direct descents of a given node
}

