% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utilities.R
\name{distEq}
\alias{distEq}
\title{Pairwise Distance from Regularly Distributed Points}
\usage{
distEq(n, prox = FALSE)
}
\arguments{
\item{n}{the number of points}

\item{prox}{a logical indicating whether to return a matrix of proximity.
Default to \code{FALSE} so the function returns the matrix of distance.}
}
\value{
A squared matrix. A matrix of class \code{dist} if \code{prox} is set to \code{FALSE}.
}
\description{
This function return a distance (or proximity) matrix between n points regularly distributed in 1 dimension.
}
\examples{
x <- distEq(5)
}
