% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utilities.R
\name{evenColors}
\alias{evenColors}
\title{Palette of evenly distributed colors}
\usage{
evenColors(n)
}
\arguments{
\item{n}{the number of colors to be in the palette.}
}
\value{
a vector of hexadecimal colors.
}
\description{
This function generates a vector of n colors evenly distributed in the RGB space.
Usefull to create palettes of distinct colors.
}
