% Generated by roxygen2 (4.0.1): do not edit by hand
\name{graphClust}
\alias{graphClust}
\title{Phylogenetically constrained  clustering}
\usage{
graphClust(p4d, trait = names(tdata(p4d)), lim.phylo = 0.2,
  lim.trait = 0.2, select.method = "ellipse", dist.phylo = "patristic",
  dist.trait = "euclidean", scale.lim = TRUE)
}
\arguments{
\item{p4d}{a phylo4d object.}

\item{trait}{the trait(s) in the phylo4d object to use for clustering.
Can be a character vector giving the name of the trait(s) or numbers giving the column index
in the table of the data slot of the p4d object.}

\item{lim.phylo}{the maximum phylogenetic distance for edges selection.}

\item{lim.trait}{the maximum traits based distance for edges selection.}

\item{dist.phylo}{a character string specifying the method used to compute phylogenetic distances.
See Details.}

\item{dist.trait}{a character string specifying the method used to compute traits distances.
See Details.}

\item{select.method}{a character string specifying the method used to select edges.}

\item{scale.lim}{logical (default TRUE) indicating if \code{lim.phylo} and \code{lim.trait} are scaled
(divided by the max value)}
}
\value{
An object of class \code{graphclust}.
}
\description{
This function extracts clusters of species based on traits values and phylogenetic proximities.
}

