% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/RcppExports.R
\name{kStarTest}
\alias{kStarTest}
\title{Computes permutation test for Blomberg's K Star}
\usage{
kStarTest(xr, vcvr, reps)
}
\arguments{
\item{xr}{a vector of data}

\item{vcvr}{phylogenetic variance-covariance matrix}

\item{reps}{the number of permutations}
}
\description{
Computes permutation test for Blomberg's K Star
}
