% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/RcppExports.R
\name{kTest}
\alias{kTest}
\title{Computes permutation test for Blomberg's K}
\usage{
kTest(xr, vcvr, reps)
}
\arguments{
\item{xr}{a vector of data}

\item{vcvr}{phylogenetic variance-covariance matrix}

\item{reps}{the number of permutations}
}
\description{
Computes permutation test for Blomberg's K
}
