% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/phyloSignal.R
\name{lambdaTest}
\alias{lambdaTest}
\title{Test Pagel's Lambda
Optimize Pagel's Lambda and do a likelihood ratio test.}
\usage{
lambdaTest(x, vcv)
}
\arguments{
\item{x}{a vector of numeric data.}

\item{vcv}{the phylogenetic variance-covariance matrix.}
}
\description{
Test Pagel's Lambda
Optimize Pagel's Lambda and do a likelihood ratio test.
}
\details{
The optimization process is currently performed in R.
Could be interesting to do this in Cpp and merge with code{pagelLogLik}.
}
