% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utilities.R
\name{.layouterize}
\alias{.layouterize}
\title{Set layout for plots}
\usage{
.layouterize(n.traits, show.tip)
}
\arguments{
\item{n.traits}{the number of traits in the layout}

\item{show.tip}{a logical indicating whether tip names are included in the layout}
}
\description{
Set layout for plots
}
