% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utilities.R
\name{.layouterizeRatio}
\alias{.layouterizeRatio}
\title{Set layout for plots}
\usage{
.layouterizeRatio(tree.ratio, n.traits, show.tip)
}
\arguments{
\item{tree.ratio}{the ratio of phylogenetic tree included in the layout}

\item{n.traits}{the number of traits in the layout}

\item{show.tip}{a logical indicating whether tip names are included in the layout}
}
\description{
Set layout for plots
}
