% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utilities.R
\name{matchTipsAndTraits}
\alias{matchTipsAndTraits}
\title{Match matrix row/col names with phylo4d tips/traits}
\usage{
matchTipsAndTraits(x, p4d = NULL, p4d.tips = NULL, p4d.traits = NULL,
  subset = TRUE)
}
\arguments{
\item{x}{a matrix of data.}

\item{p4d}{a phylo4d object.}

\item{p4d.tips}{tips labels (relevant if p4d is \code{NULL}).}

\item{p4d.traits}{traits labels (relevant if p4d is \code{NULL}).}

\item{subset}{a logical. Should the data matrix be subsetted using tips and traits labels.}
}
\value{
The data matrix (eventually subsetted).
An error if no consistency between the data and the tree.
}
\description{
Check for constitency of rows and columns names between a matrix of traits values and phylo4d object.
}
