% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/RcppExports.R
\name{moranTest}
\alias{moranTest}
\title{Computes permutation test for Moran's I}
\usage{
moranTest(xr, Wr, reps)
}
\arguments{
\item{xr}{a vector of data}

\item{Wr}{a weighting matrix}

\item{reps}{the number of permutations}
}
\description{
Computes permutation test for Moran's I
}
