% Generated by roxygen2 (4.0.1): do not edit by hand
\name{multiplot.phylo4d}
\alias{multiplot.phylo4d}
\title{Bar Plots of Traits Values along a Phylogeny}
\usage{
multiplot.phylo4d(p4d, trait = names(tdata(p4d)), center = TRUE,
  scale = TRUE, plot.type = "barplot", tree.type = "phylogram",
  tree.ratio = NULL, tree.xlim = NULL, show.tip = TRUE,
  tip.labels = NULL, tip.col = "black", tip.cex = 1, tip.font = 3,
  tip.adj = 0, bar.xlim = NULL, bar.lwd = 10, bar.col = "grey35",
  show.bar.axis = TRUE, dot.col = "black", dot.pch = 20, dot.cex = 2,
  cell.col = white2red(100), show.color.scale = TRUE, show.trait = TRUE,
  trait.labels = NULL, trait.col = "black", trait.cex = 1,
  trait.font = 1, trait.bg.col = "grey90", error.bar.sup = NULL,
  error.bar.inf = NULL, error.bar.col = 1, show.box = FALSE,
  grid.vertical = TRUE, grid.horizontal = FALSE, grid.col = "grey25",
  grid.lty = "dashed", ...)
}
\arguments{
\item{p4d}{a phylo4d object (see Details).}

\item{trait}{the trait(s) in the p4d object to include in the plot.
Can be a character vector giving the name of the trait(s) or a single number giving the column index
in the table of the data slot of the p4d object. Can be used to reorder the traits in the plot.}

\item{center}{a logical indicating whether traits values should be centered.}

\item{scale}{a logical indicating whether traits values should be scaled.}

\item{tree.type}{a character string specifying the type of phylogeny to be drawn ("phylogram" or "cladogram").}

\item{tree.ratio}{a numeric value in [0, 1] giving the proportion of width of the figure for the tree.}

\item{tree.xlim}{a numeric vector of length 2 giving the limits of the x-axis for the tree. If \code{NULL},
it is determined automatically.}

\item{show.tip}{logical indicating whether tips labels should be drawn.}

\item{tip.labels}{character vector to label the tips.
If \code{NULL} the tips labels of the \code{phylo4d} object are used}

\item{tip.col}{a vector of R colors (code or name) to use for the tips labels.
Recycled if necessary.}

\item{tip.cex}{a numeric vector to control character size of the tips labels.
Recycled if necessary.}

\item{tip.font}{an integer vector specifying the type of font for the tips labels:
1 (plain text), 2 (bold), 3 (italic), or 4 (bold italic). Recycled if necessary.}

\item{tip}{adj a vector of numeric in [0, 1] to control tips labels justification:
0 (left-justification), 0.5 (centering), or 1 (right-justification). Recycled if necessary.}

\item{bar.xlim}{numeric vector of length 2 or matrix giving the x coordinates range for
the barplots (see Details).}

\item{bar.lwd}{a vector of numeric giving bar widths of the barplot(s).
Recycled along the tips, reapeated for each trait.}

\item{bar.col}{a vector of R colors (code or name) to use for the bars.
Recycled along the tips, reapeated for each trait.
The user can also provide a matrix for a finer tuning (see Details)}

\item{show.bar.axis}{logical indicating whether barplots axes should be drawn.}

\item{dot.col}{a vector of R colors (code or name) to use for the points.
Recycled along the tips, reapeated for each trait.
The user can also provide a matrix for a finer tuning (see Details)}

\item{dot.pch}{a numerical vector of symbol to use for the points.
Recycled along the tips, reapeated for each trait.
The user can also provide a matrix for a finer tuning (see Details)}

\item{dot.cex}{a numerical vector. Character (or symbol) expansion for the points.
Recycled along the tips, reapeated for each trait.
The user can also provide a matrix for a finer tuning (see Details)}

\item{cell.col}{a list of colors for 'gridplot' cells.
Easily generated by \code{heat.colors}, \code{topo.colors}, \code{terrain.colors}
or other functions created with \code{colorRampPalette}.}

\item{show.color.scale}{logical indicating whether color scale should be drawn.}

\item{show.trait}{logical indicating whether traits labels should be drawn.}

\item{trait.labels}{character vector to label the traits.
If \code{NULL} the traits labels of the \code{phylo4d} object are used.}

\item{trait.col}{a vector of R colors (code or name) to use for the traits labels.
Recycled if necessary.}

\item{trait.cex}{a numeric vector to control character size of the trait labels.
Recycled if necessary.}

\item{trait.font}{an integer vector specifying the type of font for the traits labels:
1 (plain text), 2 (bold), 3 (italic), or 4 (bold italic). Recycled if necessary.}

\item{trait.bg.col}{a vector of R colors (code or name) to use for the background of the barplots.
Recycled if necessary.}

\item{error.bar.sup}{a matrix giving the superior limit for error bars.
Columns and rows names must match with traits and tips labels.}

\item{error.bar.inf}{a matrix giving the inferior limit for error bars.
Columns and rows names must match with traits and tips labels.}

\item{error.bar.col}{a vector of R colors (code or name) to draw error bars.}

\item{show.box}{a logical indicating whether a box should be drawn around the barplots.}

\item{grid.vertical}{a logical incating whether vertical lines of the grid should be drawn.}

\item{grid.horizontal}{a logical incating whether horizontal lines of the grid should be drawn.}

\item{grid.col}{a vector of R color(s) to use for the lines of the grid.}

\item{grid.lty}{the lines type of the grid. Possibly a vector.}

\item{further}{arguments to be passed to \code{plot.phylo}.}
}
\description{
This function computes Local Indicator of Phylogenetic Association (local Moran's I) for each tip of the tree.
Tests are based on permutations.
}

