% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/phylosignal-package.R
\docType{data}
\name{navic}
\alias{navic}
\title{Phylogeny and pollution sensitivity of diatoms}
\format{a \code{phylo4d} object.}
\source{
Keck F., Rimet F., Franc A. & Bouchez A. (In press) Phylogenetic signal in diatom ecology: perspectives for aquatic ecosystems biomonitoring. Ecological Applications.
}
\usage{
data(navic)
}
\description{
A phylogenetic tree and the pollution sensitivity of 17 diatoms species from the order Naviculales.
}
\keyword{datasets}
