% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utilities.R
\name{.orderGrArg}
\alias{.orderGrArg}
\title{Reordering vector or matrix of settings.}
\usage{
.orderGrArg(x, n.tips, n.traits, new.order, tips, default)
}
\arguments{
\item{x}{a vector or a matrix to order.}

\item{n.tips}{number of tips.}

\item{n.traits}{number of traits}

\item{new.order}{a numeric vector giving the new order.}

\item{tips}{a character vector giving the tips labels (with the new order)}

\item{default}{the default value}
}
\value{
An ordered vector or matrix.
An error if problem of consistency.
}
\description{
Internal function of multiplot.phylo4d.
Reordering vector or matrix of settings and check for names consistencies.
}
