% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/RcppExports.R
\name{pagelLogLik}
\alias{pagelLogLik}
\title{Computes log-likelihood for data and a given value of Pagel's Lambda}
\usage{
pagelLogLik(lambda, xr, vcvr)
}
\arguments{
\item{lambda}{the value of Pagel's Lambda}

\item{xr}{a vector of data}

\item{vcvr}{phylogenetic variance-covariance matrix}
}
\description{
Computes log-likelihood for data and a given value of Pagel's Lambda
}
\author{
Adapted from Liam Revell's R function 'phylosig' {phytools}.
}
