% Generated by roxygen2 (4.0.1): do not edit by hand
\name{phyloSignal}
\alias{phyloSignal}
\title{Computes phylogenetic signal with different methods}
\usage{
phyloSignal(p4d, methods = c("all", "I", "Cmean", "Lambda", "K", "K.star"),
  reps = 999)
}
\arguments{
\item{p4d}{a phylo4d object (see Details).}

\item{methods}{a character vector giving the methods to compute phylogenetic signal (see Details).}

\item{reps}{a numeric value. Number of repetitions for the estimation of p.values with randomization.}
}
\value{
A list of two dataframes with the values of statistics and associated
p.values for each tested trait and method.
}
\description{
This function computes phylogenetic signal statistics
(Blomberg's K and K*, Abouheif's Cmean, Moran's I, and Pagel's Lambda) for traits in phylo4d objects.
}
\details{
p4d has to be a phylo4d object as defined in phylobase package. phylo4d objects combine
phylogenetic tree with data. Each data associated with tips is interpreted as a trait.
By default, the \code{methods} argument is set to "all" and all the available methods are used.
The user can specify which method(s) to use. Possible values are
"\code{I}", "\code{Cmean}", "\code{Lambda}", "\code{K}" and "\code{K.star}".
}
\seealso{
\code{\link{phyloSimSignal}} .
}

