% Generated by roxygen2 (4.0.1): do not edit by hand
\name{phyloSim}
\alias{phyloSim}
\title{Simulate the behaviour of phylogenetic signal statistics with a given phylogeny}
\usage{
phyloSim(tree, methods = c("all", "I", "Cmean", "Lambda", "K", "K.star"),
  nsim = 99, reps = 999, model = "BM", pb = TRUE)
}
\arguments{
\item{tree}{a phylo, phylo4 or phylo4d object (see Details).}

\item{methods}{a character vector giving the methods to compute phylogenetic signal (see Details).}

\item{nsim}{a numeric value. Number of simulated traits at each step in the gradient.}

\item{reps}{a numeric value. Number of repetitions for the estimation of p.values with randomization.}

\item{model}{the model to use for traits simulation (only "BM", default, works).}

\item{pb}{a logical. Should a progress bar be printed? (default TRUE).}
}
\value{
An object of class \code{phylosim}.
}
\description{
This function simulates different phylogenetic signal statistics
for a given phylogenetic tree along a gradient of phylogenetic autocorrelation.
}
\details{
By default, the \code{methods} argument is set to "all" and all the available methods are used.
The user can specify which method(s) to use. Possible values are
"\code{I}", "\code{Cmean}", "\code{Lambda}", "\code{K}" and "\code{K.star}".
}
\seealso{
\code{\link{phyloSimSignal}}.
}

