% Generated by roxygen2 (4.0.1): do not edit by hand
\name{plot.graphclust}
\alias{plot.graphclust}
\title{Plot phylogenetically constrained  clustering}
\usage{
plot.graphclust(x, which = c("selection", "graph", "tree"), ask = TRUE)
}
\arguments{
\item{x}{a graphclust object as produced by graphClust.}

\item{which}{a character vector to select plots.
Must be one or more of \code{"selection"}, \code{"graph"}, \code{"tree"}.}

\item{logical}{if TRUE (default), the user is asked before each plot}
}
\description{
This function produces three plots (selectable by which):
a plot of edges selection based on phylogenetic against trait distances of taxa pairs,
a plot of the graph produced with the selected edges
and a plot of the clustered phylogenetic tree.
}

