% Generated by roxygen2 (4.1.1): do not edit by hand
% Please edit documentation in R/correlogram.R
\name{plot.phylocorrelogram}
\alias{plot.phylocorrelogram}
\title{Plot a phylogenetic correlogram}
\usage{
\method{plot}{phylocorrelogram}(x, xlab = "Phylogenetic distance",
  ylab = "Correlation", main = "Phylogenetic correlogram", ...)
}
\arguments{
\item{x}{a \code{phylocorrelogram} object.}

\item{xlab}{a label for the x axis.}

\item{ylab}{a label for the y axis.}

\item{main}{a main title for the plot}

\item{...}{other graphical parameters passed to the \code{plot} function.}
}
\description{
This function plots phylogenetic correlograms produced by \code{\link{phyloCorrelogram}}
}
\examples{
data(navic)
pc <- phyloCorrelogram(navic)
plot(pc)
}

