% Generated by roxygen2 (4.0.1): do not edit by hand
\name{plot.phylosim}
\alias{plot.phylosim}
\title{Plot phylosim object}
\usage{
plot.phylosim(x, what = c("stat", "pval"), stacked.methods = FALSE,
  quantiles = c(0.05, 0.95), col = 1:5, legend = TRUE, ...)
}
\arguments{
\item{x}{a phylosim object.}

\item{stacked.methods}{a logical. If different methods have been used, should they
be plotted on the same graphic (TRUE) or not (FALSE, default)}

\item{quantiles}{a vector of two numeric values giving the minimum and the maximum quantiles to plot.}

\item{col}{a vector of colour(s) for the different methods.}

\item{legend}{a logical. If stacked methods is set to TRUE,
should a legend be priented to differentiate the different methods?}
}
\description{
This function plots a phylosim object to visualize the behaviour of
phylogenetic signal statistics for a given phylogenetic tree
}
\seealso{
\code{\link{phyloSim}}.
}

