% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/phyloSimSignal.R
\name{plot.phylosimsignal}
\alias{plot.phylosimsignal}
\title{Plot signal estimation as a fraction of a Brownian Motion process.}
\usage{
\method{plot}{phylosimsignal}(x, methods = NULL, traits = NULL,
  stacked.methods = FALSE, stacked.traits = FALSE,
  print.quantiles = TRUE, col = 1:5, legend = TRUE, ...)
}
\arguments{
\item{x}{an object of class \code{phylosimsignal}.}

\item{methods}{a character vector giving the methods
(included in the \code{phylosimsignal} object) to plot.}

\item{traits}{a character vector giving the traits
(included in the \code{phylosimsignal} object) to plot.}

\item{stacked.methods}{If different methods have been used, should they
be plotted on the same graphic (\code{TRUE}) or not (\code{FALSE}, default).}

\item{stacked.traits}{If different traits have been used, should they
be plotted on the same graphic (\code{TRUE}) or not (\code{FALSE}, default).}

\item{print.quantiles}{logical stating whether quantiles should be plotted.}

\item{col}{a vector of colors for the different methods.}

\item{legend}{a logical. If \code{stacked.methods} is set to \code{TRUE},
should a legend be printed to differentiate the different methods?}

\item{...}{further arguments to be passed to or from other methods.}
}
\description{
Plot signal estimation as a fraction of a Brownian Motion process.
}
