% Generated by roxygen2 (4.0.1): do not edit by hand
\name{barplot.phylo4d}
\alias{barplot.phylo4d}
\alias{dotplot.phylo4d}
\alias{gridplot}
\title{Plot p4d Object}
\usage{
\method{barplot}{phylo4d}(p4d, trait = names(tdata(p4d)), center = TRUE,
  scale = TRUE, ...)

dotplot.phylo4d(p4d, trait = names(tdata(p4d)), center = TRUE,
  scale = TRUE, grid.horizontal = TRUE, grid.vertical = FALSE, ...)

gridplot(...)
}
\description{
Plot p4d Object
}

