% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/graphClust.R
\name{print.graphclust}
\alias{print.graphclust}
\title{#' Print results of phylogenetically constrained  clustering}
\usage{
\method{print}{graphclust}(x, ...)
}
\arguments{
\item{x}{a \code{graphclust} object as produced by \code{\link{graphClust}}.}

\item{...}{further arguments to be passed to or from other methods.}
}
\description{
#' Print results of phylogenetically constrained  clustering
}
