% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/phyloSimSignal.R
\name{print.phylosimsignal}
\alias{print.phylosimsignal}
\title{Print signal estimation as a fraction of a Brownian Motion process.}
\usage{
\method{print}{phylosimsignal}(x, ...)
}
\arguments{
\item{x}{an object of class \code{phylosimsignal}.}

\item{...}{further arguments to be passed to or from other methods.}
}
\description{
Print signal estimation as a fraction of a Brownian Motion process.
}
