% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/read.annotated.nexus.par.R
\name{read.annotated.nexus.par}
\alias{read.annotated.nexus.par}
\title{Convert \code{data.frame} to \code{list}.}
\usage{
read.annotated.nexus.par(trees)
}
\arguments{
\item{trees}{A \code{list} object with nexus tree names.}
}
\description{
Convert \code{data.frame} to \code{list}.
}
\examples{
my_result <- foo(iris)

}

