% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utilities.R
\name{rTraitContWeight}
\alias{rTraitContWeight}
\title{Traits Values Simulation}
\usage{
rTraitContWeight(tree, model = "BM", weight = 1, as.p4d = FALSE)
}
\arguments{
\item{tree}{an object of class "\code{phylo}.}

\item{model}{the model to use ("\code{BM}" or "\code{OU}").}

\item{weight}{a numeric vector with values ranging between 0 and 1
giving the balance between pure random evolution and the selected model.}

\item{as.p4d}{logical. Should the phylogenetic tree and the simulated data coerced to a \code{phylo4d} object?}
}
\value{
An object of class phyloSimSignal.
}
\description{
This function simulates the evolution of continuous characters along a phylogeny.
Traits values can be more or less influenced by the provided model.
}
