% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/sample_phylo_sig.R
\name{sample_phylo_sig}
\alias{sample_phylo_sig}
\title{Gets \code{tree} phylogenetic signal from trait to \code{matrix}.}
\usage{
sample_phylo_sig(trees, traits, s)
}
\arguments{
\item{trees}{A \code{list} object with nexus tree.}

\item{traits}{A \code{matrix} with traits.}

\item{s}{A \code{numeric} integer. Is the sample size.}
}
\description{
Gets \code{tree} phylogenetic signal from trait to \code{matrix}.
}
\examples{
my_result <- foo(iris)
}

