% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utilities.R
\name{subsetPhyloSimSignal}
\alias{subsetPhyloSimSignal}
\title{phyloSimSignal subsetting}
\usage{
subsetPhyloSimSignal(x, methods, traits)
}
\arguments{
\item{x}{an object of class \code{phyloSimSignal} to subset.}

\item{methods}{a character vector giving the methods to keep.}

\item{traits}{a character vector giving the traits to keep.}
}
\value{
A \code{phyloSimSignal} object
}
\description{
This function subset all the components of a phyloSimSignal object following the methods and traits provided.
}
