% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utilities.R
\name{directDescents}
\alias{directDescents}
\title{ID of direct descents}
\usage{
directDescents(phy, node)
}
\arguments{
\item{phy}{an object of class "\code{phylo}.}

\item{node}{the ID of the node.}
}
\value{
A vector of node numbers.
}
\description{
This function return the node number (ID) of the direct descents of a given node
}
