% Generated by roxygen2 (4.0.1): do not edit by hand
\name{focusTraits}
\alias{focusStop}
\alias{focusTips}
\alias{focusTraits}
\alias{focusTree}
\title{Focus on sub parts of a plot}
\usage{
focusTraits(x)

focusTree()

focusTips()

focusStop()
}
\arguments{
\item{x}{the trait to focus on.
Can be a character string giving the name of the trait or an integer giving
the number of the trait in order of appearance in the plot.}
}
\description{
These functions can be used after barplot.phylo4d to focus on
the different part of the plot and add graphical elements.
}
\details{
#' Use \code{focusTree} to focus on the phylogenetic tree, \code{focusTraits}
to focus on a given trait and \code{focusTips} to focus on the tips labels.
Use \code{focusStop} to close the editing and restore graphical settings.
For each part of the plot, the coordinate system is restored, making edition easier.
For the phylogeny, post-editing functions of the package \code{ape} like \code{nodelabels} can be used.
}

