% Generated by roxygen2 (4.0.1): do not edit by hand
\name{lipa}
\alias{lipa}
\title{Local Indicator of Phylogenetic Association (Moran's I)}
\usage{
lipa(p4d, trait = names(tdata(p4d)), reps = 999,
  alternative = "two-sided", prox.phylo = "patristic", as.p4d = FALSE)
}
\arguments{
\item{p4d}{a phylo4d object (see Details).}

\item{trait}{the trait(s) in the p4d object for which to compute LIPA.
Can be a character vector giving the name of the trait(s)
or a single number giving the column in the table of the data slot of the p4d object.}

\item{reps}{a numeric value. Number of repetitions for the estimation of p.values with randomization.}

\item{alternative}{a character string specifying the alternative hypothesis for the tests.
Must be one of \code{two-sided} (default), \code{greater} or \code{less}.}

\item{prox.phylo}{a character string specifying the method used to compute phylogenetic proximities.
See Details.}

\item{as.p4d}{logical. Should the results returned as a phylo4d object?}
}
\value{
If \code{as.p4d} is \code{FALSE} (default), the function returns a list:
\describe{
  \item{lipa}{A matrix of LIPA indices computed for each tip of the tree and each trait.}
  \item{p.value}{A matrix of p-values (tests of LIPA indices)}
  \item{reps}{Number of permutations for the tests}
  \item{alternative}{Alternative hypothesis for the tests}
}
If \code{as.p4d} is \code{TRUE}, the function returns a \code{phylo4d} object
with LIPA values as tips associated data.
}
\description{
This function computes Local Indicator of Phylogenetic Association (local Moran's I) for each tip of the tree.
Tests are based on permutations.
}

