% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/phyloSimSignal.R
\name{phyloSimSignal}
\alias{phyloSimSignal}
\title{Phylogenetic signal estimation as a fraction of a Brownian Motion process.}
\usage{
phyloSimSignal(p4d, phylosim, quantiles = c(0.05, 0.95))
}
\arguments{
\item{p4d}{a \code{phylo4d} object.}

\item{phylosim}{a \code{phylosim} object.}

\item{quantiles}{a vector of two numeric values between 0 and 1 giving the minimum
and the maximum quantiles to estimate statistics fluctuations.}
}
\value{
An object of class \code{phylosimsignal} with a \code{print} and a \code{plot} method.
}
\description{
This function crosses traits data with simulations to estimate the phylogenetic signal as a
fraction of a Brownian Motion process. This is experimental.
}
\seealso{
\code{\link{phyloSignal}}, \code{\link{phyloSim}}.
}
